+++ docs/@v1/changelog.md
@@ -7,6 +7,13 @@ toc:
 
 <!-- do-not-remove -->
 
+## 1.34.20 (2026-09-09)
+
+### Patch Changes
+
+- Updated `js-yaml` to `4.3.2`, `undici` to `6.28.0`, `redoc` to `2.5.3`, `styled-components` to `6.5.3`, and the OpenTelemetry packages to `2.11.0`.
+- Updated @redocly/openapi-core to v1.34.20.
+
 ## 1.34.19 (2026-08-10)
 
 ### Patch Changes