+++ packages/client-generator/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @redocly/client-generator
 
+## 0.4.8
+
+### Patch Changes
+
+- Updated @redocly/openapi-core to v2.52.1.
+
+## 0.4.7
+
+### Patch Changes
+
+- Updated @redocly/openapi-core to v2.52.0.
+
 ## 0.4.6
 
 ### Patch Changes