+++ packages/respect-core/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @redocly/respect-core
 
+## 2.52.1
+
+### Patch Changes
+
+- Updated @redocly/openapi-core to v2.52.1.
+
+## 2.52.0
+
+### Patch Changes
+
+- Updated @redocly/openapi-core to v2.52.0.
+
 ## 2.51.2
 
 ### Patch Changes